Newest Release Changes (Fixing Bug)
Had to transfer all files to a new project that was a .Net Framework file as opposed to a .Net Core file in Visual Studio. As a .Net Core application it built a .dll file instead of a .exe file so I had to transfer all the files. And a major bug was fixed.